% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locations.R
\name{add_locations}
\alias{add_locations}
\title{Adds location data to a SAFE data worksheet.}
\usage{
add_locations(obj, location_field = NULL, location_table = NULL,
  gazetteer_info = FALSE)
}
\arguments{
\item{obj}{An existing object of class \code{safe_data}}

\item{location_field}{The name of a location field in a \code{safedata} for 
which to add location data.}

\item{location_table}{An existing location table for a dataset, as generated by 
\code{get_locations}.}

\item{gazetteer_info}{Should all the gazetteer fields be included in the 
returned \code{sf} object. See \code{\link{load_gazetteer}} for details.}
}
\value{
A modified \code{safe_data}
}
\description{
This function adds location data to the rows of a SAFE data worksheet, 
converting it into a \code{\link{[sf]sf}} spatial features object.
}
