% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importSAFE.R
\name{addData}
\alias{addData}
\title{Add data worksheets to a \code{safedata} object}
\usage{
addData(obj, filePath = NULL)
}
\arguments{
\item{obj}{An existing object of class \code{safedata}.}

\item{filePath}{The complete path to the SAFE dataset .xlsx file. By 
default this value is set to \code{NULL} and the function attempts to 
access the \code{obj$filePath} variable (i.e. the file pointer stored in
the \code{safedata} object itself). When a \code{filePath} is supplied, 
it must point to a SAFE .xlsx file, otherwise an error is returned.}
}
\value{
A modified \code{safedata} object with data worksheets added.
}
\description{
This function processes data-containing worksheets for the SAFE dataset at
the specified \code{filePath} and adds them to an existing \code{safedata}
object. \code{addData} creates a new dataframe within the \code{safedata}
object for each worksheet stored in the SAFE dataset. The dataframes are
named as per those used in the SAFE dataset submission file. All data
columns are imported according to the \code{field_type} provided in the
SAFE data table. Header information for each worksheet is stored separately
as a \code{metaInfo} attribute accessed via the dataframe.
}
\examples{
  # create a safedata object and add data
  filePath <- system.file('extdata', 'demo_data.xlsx', package='safedata')
  safe <- importSafe(filePath)
  safe <- addData(safe, filePath)
  
  # access the data table named "data_1"
  View(safe$data_1)
  
  # access the metadata for this table
  View(attr(safe$data_1, 'metaInfo'))
}
\seealso{
\url{https://safe-dataset-checker.readthedocs.io/en/latest/data_format/data/}
  for information on SAFE data worksheets
}
