% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importSAFE.R
\name{addLocations}
\alias{addLocations}
\title{Add location information to a \code{safedata} object}
\usage{
addLocations(obj, filePath = NULL)
}
\arguments{
\item{obj}{An existing object of class \code{safedata}}

\item{filePath}{The complete path to the SAFE dataset .xlsx file. By 
default this value is set to \code{NULL} and the function attempts to 
access the \code{obj$filePath} variable (i.e. the file pointer stored in
the \code{safedata} object itself). When a \code{filePath} is supplied, 
it must point to a SAFE .xlsx file, otherwise an error is returned.}
}
\value{
A modified \code{safedata} object with \code{Locations} added
}
\description{
Processes the Locations worksheet in the SAFE dataset at the supplied
\code{filePath} and adds it as a dataframe to an existing \code{safedata}
object (\code{obj}). When no Locations worksheet is found the function
defaults to \code{NA}. The Locations worksheet contains details of the
commonly used areas in which research was conducted at SAFE. In some cases
locations will already be "known" (i.e. on the SAFE gazetteer at
\url{https://www.safeproject.net/info/gazetteer}), in other cases they will
be "new", and thus need to be accompanied by GPS coordinates.
}
\note{
Although unusual, not all SAFE project submissions will contain a
  Locations worksheet. Examples of this include projects that present only
  laboratory data (that do not have the requirement of specifying where
  field samples came from), or for inconsistent data collection locations
  (e.g. tracking animal movements). In the case of the latter GPS data
  should be provided separately for each observation.
}
\seealso{
\url{https://safe-dataset-checker.readthedocs.io/en/latest/data_format/locations/}
  for details on the Locations worksheet,
  \url{https://www.safeproject.net/info/gazetteer} for the SAFE gazetteer
}
