% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importSAFE.R
\name{addTaxa}
\alias{addTaxa}
\title{Add taxonomic observations to a \code{safedata} object}
\usage{
addTaxa(obj, filePath = NULL)
}
\arguments{
\item{obj}{An existing object of class \code{safedata}}

\item{filePath}{The complete path to the SAFE dataset .xlsx file. By 
default this value is set to \code{NULL} and the function attempts to 
access the \code{obj$filePath} variable (i.e. the file pointer stored in
the \code{safedata} object itself). When a \code{filePath} is supplied, 
it must point to a SAFE .xlsx file, otherwise an error is returned.}
}
\value{
A modified \code{safedata} object with a \code{Taxa} dataframe
}
\description{
This function adds taxonomic data from the Taxa worksheet in a SAFE project
dataset - in the form of a table - to an existing \code{safedata} object.
If no Taxa worksheet is provided with the dataset, the table defaults to
\code{NA}.
}
\note{
Not all SAFE project submissions contain the Taxa worksheet (for
  example if the data do not contain any taxonomic observations). In this
  case \code{safedata$Taxa} defaults to \code{NA}.
}
\seealso{
\code{\link{addTaxonHeirarchies}},
  \url{https://safe-dataset-checker.readthedocs.io/en/latest/data_format/taxa/}
  for information on the Taxa worksheet, \code{\link{addTaxonHeirarchies}}
}
