% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importSAFE.R
\name{addTaxonHeirarchies}
\alias{addTaxonHeirarchies}
\title{Add complete taxonomic heirarchies to a \code{safedata} object}
\usage{
addTaxonHeirarchies(obj, ...)
}
\arguments{
\item{obj}{An existing object of class \code{safedata}}

\item{...}{Optional arguments to pass to \code{\link[rgbif]{name_backbone}}}
}
\value{
The updated \code{safedata} object with a dataframe of taxonomic
  heirarchies for all taxa listed in the Taxa worksheet of the SAFE dataset.
  This is accessed using the reference \code{TaxonHeirarchy}
}
\description{
Adds a dataframe of taxonomic heirarchies for all taxa reported in the SAFE
dataset Taxa worksheet to the supplied \code{safedata} object. When no
Taxa worksheet exists, this function flags a warning.
}
\note{
Not all SAFE datasets contain the Taxa worksheet. In this case, a
  value of \code{NA} will be assigned to the \code{TaxonHeirarchy}
}
\seealso{
\code{\link{addTaxa}}, \code{\link{getNameBackbone}}
}
