% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchSAFE.R
\name{apiSearch}
\alias{apiSearch}
\title{Run an API call using the constructed \code{api} URL.}
\usage{
apiSearch(api, ids = NULL)
}
\arguments{
\item{api}{The constructed Zenodo SAFE database search API.}

\item{ids}{A set of (numeric) Zenodo record IDs that restricts the returned
search values. Only record IDs identified in the search API that appear
in the \code{ids} subset are returned to the user.}
}
\value{
A set of numeric Zenodo record IDs for SAFE datasets found to match
  the search criteria specified in the \code{api} construct.
}
\description{
\code{apiSearch} is a wrapper function that downloads a URL (in this case
assumed to be a Zenodo SAFE database search API construct) and converts the
resulting JSON data into a readable R object (a \code{list}). An array of
Zenodo record IDs (matching the Zenodo search API results) are extracted
and compared against a supplied set of (optional) \code{ids}, enabling the
user to restrict the search outcome to a particular set of record IDs.
}
\note{
It is assumed that the \code{api} URL passed is of the correct format
  to perform a search of the Zenodo SAFE database.
}
\seealso{
\code{\link{searchDates}}, \code{\link{searchFields}},
  \code{\link{searchAuthor}}, \code{\link{searchTaxa}},
  \code{\link{searchText}}, \code{\link{searchSpatial}} for building
  different search query types, \url{https://www.safeproject.net/api/} for
  a full outline of the Zenodo search API used to interface with SAFE
  projects.
}
