% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSAFE.R
\name{buildFilePathFromVersions}
\alias{buildFilePathFromVersions}
\title{Builds a list of file paths to SAFE datasets in the versions list}
\usage{
buildFilePathFromVersions(versions)
}
\arguments{
\item{versions}{list of all record versions contained within a given
concept ID. The output from a call to \code{\link{zenodoVersionsApiLookup}}}
}
\value{
a list object containing (prospective) file paths to the data files
  associated with each record within the concept ID, constructed according
  to the \code{safedata} directory structure
}
\description{
Using the output from a call to the Zenodo "versions" API
(\code{\link{zenodoVersionsApiLookup}}), this function builds a set of
pathways to local copies of downloaded datasets within a given SAFE concept
ID. The structure of these pathways is governed by the \code{safedata}
directory protocol, i.e. concept_ID > record_ID > file_name.
}
\seealso{
\code{\link{zenodoVersionsApiLookup}}
}
