% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSAFE.R
\name{buildVersionsDataframe}
\alias{buildVersionsDataframe}
\title{Build a "versions" dataframe of a SAFE concept ID}
\usage{
buildVersionsDataframe(allVersions)
}
\arguments{
\item{allVersions}{the full versions list returned from a call to 
\code{\link{zenodoVersionsApiLookup}}}
}
\value{
a dataframe containing various information for each record version
  associated with a particular SAFE concept ID
}
\description{
Takes the output from a call to the Zenodo 'versions' API and processes the
returned \code{list} into a dataframe storing version details for all
records associated with a given SAFE project (identified using the concept
record ID number). Version fields include the DOI, record ID, creation
date, access status ('open', 'embargoed', or 'closed'), and - in the case 
of embargoed records - the open date (the date that the record will be
publicly accessible) for all records associated with the concept ID.
}
\seealso{
\code{\link{zenodoVersionsApiLookup}}, \code{\link{updateIndex}}
}
