% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{cache_index}
\alias{cache_index}
\title{Load the record index into memory}
\usage{
cache_index()
}
\description{
This function loads the dataset record index from the JSON
file in the SAFE data directory and sets which datasets are
currently available and which records are the most recent 
available under a given concept.
}
\keyword{internal}
