% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchSAFE.R
\name{checkApiCall}
\alias{checkApiCall}
\title{Check a SAFE search API call for errors.}
\usage{
checkApiCall(apiOutput)
}
\arguments{
\item{apiOutput}{The output of a call to the Zenodo SAFE database search
API. This should be a \code{list} object.}
}
\description{
\code{checkApiCall} is a simple utility function used to check calls to the
Zenodo SAFE database search API for invalid results. \code{checkApiCall}
checks for a 404 message on the search output and prints the returned error
message for user information.
}
\note{
This function is for background use only, e.g. in functions such as
  \code{\link{searchDates}}, \code{\link{searchTaxa}}, to verify API calls.
}
