% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSAFE.R
\name{checkRecordDownloaded}
\alias{checkRecordDownloaded}
\title{Check if the record (file) at the specified path has been downloaded}
\usage{
checkRecordDownloaded(path)
}
\arguments{
\item{path}{the full path to the prospective file}
}
\value{
\code{TRUE}/\code{FALSE} value indicating whether file was found at
  the specified path
}
\description{
Check if the record (file) at the specified path has been downloaded
}
