% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importSAFE.R
\name{createSafe}
\alias{createSafe}
\title{Initiate a new \code{safedata} object}
\usage{
createSafe(filePath = NULL, processSummary = TRUE)
}
\arguments{
\item{filePath}{the complete path to the SAFE dataset .xlsx file. This 
could be absolute, or relative to the current working directory. When no
path is specified (i.e. \code{filePath = NULL}), an empty \code{safedata}
object is returned.}

\item{processSummary}{a logical value indicating whether summary data
should be added to the \code{safedata} object on creation. If \code{TRUE},
then summary metadata is added to the object using the function
\code{\link{processSafeSummary}}.}
}
\value{
an object of class \code{safedata}.
}
\description{
\code{create_safe} generates an object of class type \code{safedata} that
is used to import and view SAFE project datasets.
}
\seealso{
\code{\link{processSafeSummary}}
}
