% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{download_gazetteer}
\alias{download_gazetteer}
\title{Downloads the current SAFE gazetteer}
\usage{
download_gazetteer()
}
\description{
This function downloads the gazetteer from the SAFE project
gazetteer API (\url{https://www.safeproject.net/api/gazetteer})
and saves it in the root of the safe data directory.
}
\keyword{internal}
