% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{download_index}
\alias{download_index}
\title{Download the current dataset index}
\usage{
download_index()
}
\description{
This function downloads the dataset index from the SAFE project
index API (\url{https://www.safeproject.net/api/index}) and saves
it in the root SAFE data directory
}
\keyword{internal}
