% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{download_location_aliases}
\alias{download_location_aliases}
\title{Downloads the current SAFE location aliases}
\usage{
download_location_aliases()
}
\description{
This function downloads the location aliases from the SAFE data
location aliases API (for example,
\url{https://www.safeproject.net/api/location_aliases})
and saves it in the root of the safe data directory.
}
\keyword{internal}
