% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSAFE.R
\name{downloadSafe}
\alias{downloadSafe}
\title{Download SAFE project data file associated with the supplied record}
\usage{
downloadSafe(zenodoRecord, dir = NULL)
}
\arguments{
\item{zenodoRecord}{\code{list} object returned from Zenodo database API
call. This should be a call to the desired SAFE project}

\item{dir}{directory into which data file should be downloaded}
}
\description{
Attempts to download the data file associated with the supplied Zenodo
record, which should be a part of the SAFE community. The file will be
downloaded into the specified directory (\code{dir}). \code{safedata}
version 0.1 currently supports only .xlsx files. Future releases may be
expanded to enable downloading of other file types.
}
\seealso{
\code{\link{zenodoRecordApiLookup}}, \code{\link{setSafeDir}}
}
