% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{fetch_record_metadata}
\alias{fetch_record_metadata}
\title{Get SAFE dataset metadata}
\usage{
fetch_record_metadata(record_set)
}
\arguments{
\item{record_set}{An object of class \code{safe_record_set}.}
}
\description{
Internal handler to ensure there are local copies of record metadata,
fetching it from the SAFE project website if needed. This is the same
data used to populate the Zenodo description but is machine readable
and contains additional taxon and location indexing.
}
\examples{
  recs <- c('https://doi.org/10.5281/zenodo.3247631', '10.5281/zenodo.3266827', 
           'https://zenodo.org/record/3266821', 'not_this_one/3266821')
  recs <- validate_record_ids(recs)
  fetch_record_metadata(recs)
}
\keyword{internal}
