% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{file_status}
\alias{file_status}
\title{Get file status for a dataset}
\usage{
file_status(metadata)
}
\arguments{
\item{A}{loaded metadata dictionary for a record}
}
\description{
This returns a formatted status string for \code{\link{show_record}} and
\code{\link{show_worksheet}} showing if a dataset is available globally
or locally as a private copy.
}
\keyword{internal}
