% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_safe.R
\name{get_field_metadata}
\alias{get_field_metadata}
\title{Extract the field metadata for a safedata object.}
\usage{
get_field_metadata(obj)
}
\arguments{
\item{obj}{A safedata object}
}
\description{
This function returns a data frame containing the field metadata from 
a safedata object. This is a dataframe with field descriptors as columns
and a row for each field in the safedata object. The function will add
blank metadata rows for fields that have been added to the dataset and
will subset metadata to ensure that the rows of the metadata align with 
the columns of the safedata object.
}
\details{
Note that this function currently relies on matching field names, so 
changing the column names of safedata will break it.
}
\examples{
   set_example_safe_dir()
   beetle_abund <- load_safe_data(1400562, 'Ant-Psel')
   get_field_metadata(beetle_abund)
   unset_example_safe_dir()
}
