% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa.R
\name{get_field_taxa}
\alias{get_field_taxa}
\title{Extract taxa associated with fields}
\usage{
get_field_taxa(obj, taxon_table = NULL, which = NULL, na.values = NULL)
}
\arguments{
\item{obj}{An existing object of class \code{safedata}}

\item{taxon_table}{An existing taxon table for a dataset, as generated by 
\code{\link{get_taxa}}.}

\item{which}{A vector specifying a subset of taxonomy table field names to add.}

\item{na.values}{A vector of values that count as NAs in field metadata.}
}
\value{
A modified \code{safedata} object with added taxonomic columns.
}
\description{
A \code{safedata} object can contain field metadata identifying the
taxa sampled for a particular field. Often, taxon identities are stored
in a separate field (of type 'taxon'), but sometimes all values in a field
are from a constant set of taxa. These could show that a field contains 
the abundance of a particular species, or trait values for a single species
or interaction data between specific species.
}
\details{
This function extracts these taxon identities and returns a data frame 
containing the taxon information for those fields. The data frame will
have one row for each taxon identity in each field, so interaction fields
may have multiple rows.
}
\note{
This function looks for fields that have non-NA values in the
   taxon_name and interaction_name field descriptors. These descriptors
   should be NA when unused, but some datasets have used other null values
   (e.g. None). In these cases, a warning about unknown taxa will be issued
   and the \code{na.values} argument can then be used to exclude these values.
}
\examples{
   set_example_safe_dir()
   beetle_morph <- load_safe_data(1400562, 'MorphFunctTraits')
   beetle_morph <- add_taxa(beetle_morph)
   unset_example_safe_dir()
}
\seealso{
\code{\link{get_taxa}}
}
