% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{get_index}
\alias{get_index}
\title{Get the cached dataset index}
\usage{
get_index()
}
\value{
A data frame containing the dataset index details.
}
\description{
This function just safely retrieves the safedata index from the
package cache.
}
\seealso{
\code{\link{load_index}}
}
\keyword{internal}
