% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa.R
\name{get_taxa}
\alias{get_taxa}
\title{Obtaining taxonomy for a SAFE dataset.}
\usage{
get_taxa(obj)
}
\arguments{
\item{obj}{A single record id, or an existing safedata dataframe.}
}
\value{
A taxonomy table of classes "safe_taxa" and "data.frame".
}
\description{
This function generates a taxonomy table for a specified SAFE dataset
record. Each row represent a taxon used in the data worksheets in the
dataset and the table fields show the taxonomic hierarchy for those taxa.
}
\details{
All SAFE datasets containing taxa must include worksheets providing
taxonomy details, used to validate the taxa in the dataset against
either or both of the GBIF or NCBI taxonomy database. All entries in
these worksheet are  validated against the databases before
publication and are used to populate a taxonomic index for safedata
datasets. The validated taxonomic index for a specific dataset is
also available in the metadata for a record and this function converts
that metadata into a taxonomy table for the dataset.

The returned table includes the fields \code{worksheet_name},
\code{taxon_name}, \code{taxon_rank} and \code{taxon_status}, along
with the taxonomy database used to validate the taxon
(\code{taxon_auth}). These fields provide the original taxonomy
worksheet data included.


For more details on the structure of the taxon worksheets see:
\url{https://safedata-validator.readthedocs.io/en/latest/data_format/taxa.html}
}
\section{Synonyms}{

The taxon table will include extra rows for worksheet taxa that are
GBIF synonyms or are merged in the NCBI taxonomy. There will be a row
giving the taxonomic hierarchy for the original taxon name provided
with the dataset and an additional row providing the canonical taxon
name. The worksheet name for these rows will be identical.
}

\section{Taxonomic ranks}{

NCBI and GBIF differ in their use of taxonomic ranks: GBIF uses a
reduced set of core 'backbone' ranks, where NCBI supports a wider
range of ranks. The taxon hierarchy used for safedata reduces NCBI to
the core GBIF backbone ranks but adds Superkingdom, which is used for
bacterial and viral groups.

Especially note that the taxonomy used by GBIF and NCBI are
**not congruent**: if the same taxon is included using both systems,
it may well have substantially different ranks.
}

\examples{
    set_example_safe_dir()
    taxa <- get_taxa(1400562)
    unset_example_safe_dir()
}
\seealso{
\code{\link{add_taxa}}
}
