% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importSAFE.R
\name{getNameBackbone}
\alias{getNameBackbone}
\title{Return taxonomic heirarchy for a SAFE taxonomic entry}
\usage{
getNameBackbone(taxaRow, ...)
}
\arguments{
\item{taxaRow}{A row from a SAFE object Taxa worksheet}

\item{...}{Optional arguments to pass to \code{\link[rgbif]{name_backbone}}}
}
\value{
GBIF name backbone list
}
\description{
This function attempts to perform a search of the GBIF database for the
named Taxon/Parent types in the SAFE project Taxa worksheet. The purpose
of this is to construct a taxonomic tree for all taxa reported in the
dataset(s) to facilitate data searching/subsetting.
}
\note{
This function can take a few seconds to run, particularly if there
  are a large number of Taxa in the SAFE dataset.
}
\seealso{
\code{\link[rgbif]{name_backbone}},
  \code{\link[rgbif]{name_usage}},
  \url{https://safe-dataset-checker.readthedocs.io/en/latest/data_format/taxa/},
  \url{https://safe-dataset-checker.readthedocs.io/en/latest/safedataset_checker/gbif_validation/}
}
