% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSAFE.R
\name{getRecordIdFromDoi}
\alias{getRecordIdFromDoi}
\title{Extracts record ID from Zenodo DOI link}
\usage{
getRecordIdFromDoi(DOI)
}
\arguments{
\item{DOI}{DOI link for a Zenodo-stored project}
}
\value{
The record ID component associated with the DOI, i.e. the numeric
  components that appear after the final "/"
}
\description{
Extracts record ID from Zenodo DOI link
}
