% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSAFE.R
\name{getSafe}
\alias{getSafe}
\title{Get the SAFE files with the specified IDs}
\usage{
getSafe(ids, dir = NULL, ...)
}
\arguments{
\item{ids}{an individual or array of project ID numbers to be downloaded}

\item{dir}{SAFE data directory, which defaults to the \code{SAFE_data_dir} 
option. A new directory is created for each concept record ID. Data for
individual records are stored within the concept record ID directory}

\item{...}{optional argument to be passed to \code{\link{processSafe}}}
}
\description{
This function processes and downloads the data files associated with the
SAFE project IDs supplied. Each ID (which may be either a concept record ID
or specific record ID) is validated as being from a SAFE community project,
checked for versioning and access permissions, and finally downloaded into
the relevant folder. Lastly, the ID downloaded statuses are updated in the
SAFE index.rds file.
}
\examples{
  getSafe(3081059, "C:/Users/User/Desktop/SAFE_data/")
  getSafe(c(3081059, 2537074), "C:/Users/User/Desktop/SAFE_data/")
}
\seealso{
\code{\link{processSafe}}, \code{\link{downloadSafe}}
}
