% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importSAFE.R
\name{getSafeDataClassType}
\alias{getSafeDataClassType}
\title{Get the R data type from the SAFE \code{field_type} variable}
\usage{
getSafeDataClassType(safeType)
}
\arguments{
\item{safeType}{The SAFE \code{field_type} variable}
}
\value{
The corresponding readxl data type
}
\description{
Takes a SAFE \code{field_type} variable and returns the appropriate ReadXl
data type. This function is handy for ensuring SAFE worksheet data tables
are correctly imported by ReadXl (i.e. that format types are consistent).
If the data type is not identifiable then this function will return a value
of "guess".
}
\seealso{
\url{https://safe-dataset-checker.readthedocs.io/en/latest/data_format/data/}
  for accepted SAFE data field types and 
  \url{https://cran.r-project.org/web/packages/readxl/} for ReadXl data types
}
