% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importSAFE.R
\name{getTaxonWrapper}
\alias{getTaxonWrapper}
\title{Extract taxonomic heirarchy from the GBIF database for the given taxa}
\usage{
getTaxonWrapper(taxaRow, ...)
}
\arguments{
\item{taxaRow}{A row from a SAFE object Taxa worksheet}

\item{...}{Optional arguments to pass to \code{\link[rgbif]{name_backbone}}}
}
\value{
Dataframe containing the taxonomic heirarchy for the given taxon
}
\description{
This function is a wrapper that provides capabilities to extract taxonomic
herirarchies from the GBIF database using Taxa worksheet in a SAFE data 
file.
}
\seealso{
\code{\link{getNameBackbone}}, \code{\link{nameBackboneToDf}}
}
