% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importSAFE.R
\name{importSAFE}
\alias{importSAFE}
\title{Import a SAFE data file}
\usage{
importSAFE(file)
}
\arguments{
\item{file}{Path to the .xlsx SAFE file}
}
\value{
A SAFE object \code{list} with varied attributes, including
  \code{Summary}, \code{Taxa}, \code{Locations}, and individual data
  fields for accessing different components of the SAFE record data.
}
\description{
This function is a wrapper to open and process SAFE data tables stored in
memory as .xlsx files. A full path to a SAFE .xlsx file should be provided,
which is then opened and processed according to the following procedure:
1. Process \code{Summary} worksheet
2. Process \code{Taxa} worksheet (when provided)
3. Process \code{Locations} worksheet (when provided)
4. Process individual data worksheets
More information on SAFE .xlsx file structures can be found at
\url{https://safe-dataset-checker.readthedocs.io/en/latest/data_format/overview/#excel-format-overview}.
}
\seealso{
\code{\link{getSAFE}} for downloading SAFE files from the Zenodo
  cloud database, \url{https://safe-dataset-checker.readthedocs.io/en/latest/data_format/overview/#excel-format-overview}
  for an overview on Excel file formats used for SAFE data submissions.
}
