% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importSAFE.R
\name{isCategorical}
\alias{isCategorical}
\title{Check if the SAFE data type is of R type "categorical"/"factor"}
\usage{
isCategorical(safeType)
}
\arguments{
\item{safeType}{The SAFE \code{field_type} variable}
}
\value{
\code{TRUE} if given variable is a categorical, \code{FALSE} if not
}
\description{
Takes a SAFE \code{field_type} and checks whether it is of R type
"categorical"/"factor". This is used to convert SAFE data variables
after import.
}
\seealso{
\code{\link{getDataClass}},
  \url{https://safe-dataset-checker.readthedocs.io/en/latest/data_format/data/}
  for differernt SAFE data types
}
