% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSAFE.R
\name{isSAFE}
\alias{isSAFE}
\title{Checks whether the supplied zenodo record is from the SAFE community}
\usage{
isSAFE(zenodoRecord)
}
\arguments{
\item{zenodoRecord}{returned \code{list} object from an API call to the
Zenodo cloud storage database. Should contain information pertaining to
a Zenodo record (or records)}
}
\value{
\code{TRUE}/\code{FALSE} indicating whether the record is part of
  the SAFE community
}
\description{
This function returns a boolean indicating whether the supplied Zenodo
record information (the result of an API lookup to the Zenodo database) is 
part of the SAFE (Stability of Altered Forest Ecosystems) community.
}
\seealso{
\url{https://zenodo.org/communities/safe/},
  \code{\link{zenodoRecordApiLookup}}, \code{\link{zenodoVersionsApiLookup}}
}
