% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{load_location_aliases}
\alias{load_location_aliases}
\title{Load and cache the SAFE location aliases}
\usage{
load_location_aliases()
}
\value{
A data frame containing the SAFE location aliases.
}
\description{
This function loads the SAFE locations alias, stored as a csv file in
the root of the SAFE data directory, as data frame.
}
\details{
When this function is first called in a session, the loaded data frame
object is cached for re-use (see \code{\link{load_index}}).
}
\seealso{
\code{\link{load_gazetteer}}, \code{\link{load_index}}
}
\keyword{internal}
