% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{load_record_metadata}
\alias{load_record_metadata}
\title{Loads the metadata for a record}
\usage{
load_record_metadata(record_set)
}
\arguments{
\item{record_set}{An object of class \code{safe_record_set} containing a single
row with complete concept and record data.}
}
\description{
Loads the metadata for a record
}
\keyword{internal}
