% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_safe.R
\name{load_safe_worksheet}
\alias{load_safe_worksheet}
\title{Internal data loading function}
\usage{
load_safe_worksheet(record_set, worksheet, skip, n_max)
}
\arguments{
\item{worksheet}{The name of the worksheet to load (str).}

\item{skip}{The number of initial rows to skip (integer)}

\item{n_max}{The number of rows to load (integer)}

\item{index_row}{A row selected from the index data frame, identifying the
workbook path.}
}
\value{
A data frame of the worksheet,
}
\description{
This is an internal function shared between load_safe_data (data worksheets)
get_taxa (Taxa sheets) and get_locations (Locations sheets).
}
\keyword{internal}
