% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_mockr.R
\docType{data}
\name{mock_api_urls}
\alias{mock_api_urls}
\title{URLs handled by mock_api}
\format{An object of class \code{list} of length 20.}
\usage{
mock_api_urls
}
\description{
This list provides a set of URLs handled by the \code{mock_api} URL
mocking function and handler. The list provides the matched URL along with
the name of the file (in api \code{inst/api_data} directory) used to
provide any content in the response body and the status code to be
returned.
}
\keyword{internal}
