% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importSAFE.R
\name{nameBackboneToDf}
\alias{nameBackboneToDf}
\title{Convert RGBIF name_backbone list to dataframe}
\usage{
nameBackboneToDf(nameBackbone)
}
\arguments{
\item{nameBackbone}{List of taxonomic heirarchy returned by 
\code{\link[rgbif]{name_backbone}}}
}
\value{
Taxonomic heirarchy stored as a dataframe including the 8 main
  taxonomic levels
}
\description{
For improved readability and searchability, this function converts an
\code{rgbif} name backbone (list) to a dataframe, storing information for
the 8 main taxonomic levels (subspecies, species, genus, family, class,
order, phylum, kingdom). The dataframe also includes the unique taxon
identifier within the SAFE project submission and the \code{matchtype}
variable returned from \code{rgbif}.
}
\seealso{
\code{\link{getTaxonHeirarchy}}, \code{\link[rgbif]{name_backbone}}
}
