% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_safe.R
\name{parse_factor_metadata}
\alias{parse_factor_metadata}
\title{Parses the factor metadata format into a dataframe}
\usage{
parse_factor_metadata(x)
}
\arguments{
\item{x}{A string containing factor information in the metadata format.}
}
\value{
A data frame of factor levels and description (if any)
}
\description{
Parses the factor metadata format into a dataframe
}
\keyword{internal}
