% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importSAFE.R
\name{print.safedata}
\alias{print.safedata}
\title{Print summary of information for a \code{safedata} object to the console}
\usage{
\method{print}{safedata}(x, ...)
}
\arguments{
\item{x}{An object of class \code{safedata}}

\item{...}{Further arguments to print generic, ignored}
}
\description{
Prints a summary of metadata for the given SAFE project object to the
command line. Includes the project title, project ID number, start and end
dates, and data worksheet names.
}
\seealso{
\code{\link{createSafe}}, \code{\link{processSafeSummary}}
}
