% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importSAFE.R
\name{printSummary}
\alias{printSummary}
\title{Prints summary information for SAFE object}
\usage{
printSummary(safeObj)
}
\arguments{
\item{safeObj}{Existing SAFE data object with meta information added}
}
\description{
Prints a summary of metadata for the given SAFE project object to the
command line. Includes the project title, project ID number, start and end
dates, and data worksheet names.
}
\seealso{
\code{\link{processSummary}}
}
