% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSAFE.R
\name{printVersions}
\alias{printVersions}
\title{Returns print-friendly table displaying versions information for a given
SAFE concept ID}
\usage{
printVersions(printVers)
}
\arguments{
\item{printVers}{dataframe of versions information}
}
\value{
print-friendly table in the form of a string
}
\description{
Returns print-friendly table displaying versions information for a given
SAFE concept ID
}
\seealso{
\code{\link{processSafe}}
}
