% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importSAFE.R
\name{processData}
\alias{processData}
\title{Add data worksheets to SAFE project object}
\usage{
processData(file, safeObj)
}
\arguments{
\item{file}{Path to the SAFE project file. This is assumed to be .xlsx}

\item{safeObj}{An existing SAFE data object}
}
\value{
A modified SAFE data object containing data worksheets
}
\description{
This function processes data-containing worksheets for the given SAFE
project file and adds them to an existing SAFE data object. This function
works by creating a new \code{dataframe} within the SAFE object for each
worksheet found within the SAFE project file. The dataframes are named
according to their names in the original SAFE project file and contain the
data within the given workseet. All data columns are imported according to 
the \code{field_type} provided in the SAFE data table. Header information
for each worksheet is stored as a \code{metaInfo} attribute accessed via
the dataframe.
}
\note{
To access worksheet meta information in a SAFE object \code{safeObj}
  with worksheet \code{data}, it is recommended to use
  \code{View(attr(safeObj$data, 'metaInfo'))}
}
\seealso{
\url{https://safe-dataset-checker.readthedocs.io/en/latest/data_format/data/}
  for information on SAFE data worksheets
}
