% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importSAFE.R
\name{processLocations}
\alias{processLocations}
\title{Add location information to SAFE data object}
\usage{
processLocations(file, safeObj)
}
\arguments{
\item{file}{Complete path to the SAFE object file}

\item{safeObj}{An existing SAFE data object}
}
\value{
A modified SAFE object with a \code{Locations} dataframe
}
\description{
Processes the \code{Locations} worksheet in the SAFE project file and adds
it as a dataframe to the existing SAFE data object.
}
\note{
Although unusual, not all SAFE project submissions will contain a
  Locations worksheet. Examples of this include projects that present only
  laboratory data (that do not have the requirement of specifying where
  field samples came from), or for inconsistent data collection locations
  (e.g. tracking animal movements). In the case of the latter GPS data
  should be provided separately for each observation.
}
\seealso{
\url{https://safe-dataset-checker.readthedocs.io/en/latest/data_format/locations/}
  for details on the Locations worksheet
}
