% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importSAFE.R
\name{processSafeSummary}
\alias{processSafeSummary}
\title{Process summary 'meta' information for SAFE project data}
\usage{
processSafeSummary(obj, filePath = NULL)
}
\arguments{
\item{obj}{An object of class \code{safedata}.}

\item{filePath}{The complete path to the SAFE dataset .xlsx file. By default this
value is set to \code{NULL} and the function attempts to access the
\code{obj$filePath} variable (i.e. the file pointer stored in the
\code{safedata} object itself). When a \code{filePath} is supplied, it
must point to a SAFE .xlsx file, otherwise an error is returned.}
}
\value{
\code{safedata} object with metadata added.
}
\description{
Adds summary information to a \code{safedata} object using the dataframe
provided. \code{processSafeSummary} reads and assigns variables from the
summary dataframe - a transposed "Summary" worksheet from a SAFE project
.xlsx dataset - into a \code{safedata} object.
}
\seealso{
\url{https://safe-dataset-checker.readthedocs.io/en/latest/data_format/summary/}
  for information on the SAFE summary worksheet, \code{\link{createSafe}}
}
