% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importSAFE.R
\name{processSummary}
\alias{processSummary}
\title{Process summary 'meta' information for SAFE project data}
\usage{
processSummary(df)
}
\arguments{
\item{dataframe}{Dataframe containing summary information. This must be a 
"Summary" worksheet from a SAFE project submission file}
}
\value{
SAFE List object containing summary information for the project
}
\description{
Initiates a SAFE data object from the summary sheet provided. This function
reads and assigns variables from the provided summary dataframe (a
transposed "Summary" worksheet from a SAFE project submission file) into a
new SAFE \code{list} object.
}
\seealso{
\url{https://safe-dataset-checker.readthedocs.io/en/latest/data_format/summary/}
  for information on the SAFE summary worksheet
}
