% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importSAFE.R
\name{processTaxa}
\alias{processTaxa}
\title{Add taxa information to SAFE data object}
\usage{
processTaxa(file, safeObj)
}
\arguments{
\item{file}{Complete path to the SAFE project file}

\item{safeObj}{An existing SAFE data object}
}
\value{
A modified SAFE object with a \code{Taxa} dataframe
}
\description{
This function adds a new dataframe to an existing SAFE object containing
all taxonomic information for the submitted dataset. Note that when the
data do not contain any taxa this table defaults to \code{NA}. When
provided, this dataframe is used to compile a complete taxonomic heirarchy 
for all taxa identified within the dataset.
}
\note{
Not all SAFE project submissions contain the Taxa worksheet (for
  example if the data do not contain taxa). In this case the SAFE Taxa
  object defaults to \code{NA}.
}
\seealso{
\code{\link{buildTaxonHeirarchy}},
  \url{https://safe-dataset-checker.readthedocs.io/en/latest/data_format/taxa/}
  for information on the Taxa worksheet
}
