% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{readTransposedXlsx}
\alias{readTransposedXlsx}
\title{Read a transposed .xlsx file into a dataframe}
\usage{
readTransposedXlsx(path, sheetName, ...)
}
\arguments{
\item{path}{The path to the .xlsx file to be opened}

\item{sheetName}{The name of the worksheet to be imported}

\item{...}{Optional arguments to be passed to \code{read_xlsx}}
}
\value{
The reformatted .xlsx with headers as columns and data in rows
}
\description{
Provides capability to read .xlsx files that are transposed (i.e. 
organised with headers in rows and data across columns) into a standard R
dataframe. This function preserves data types.
}
\seealso{
\code{\link[readxl]{read_xlsx}}
}
