% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{retrieve_index}
\alias{retrieve_index}
\title{Get the cached SAFE dataset index}
\usage{
retrieve_index()
}
\description{
Internal handler to retrieve the local loaded copy of the index from the
package cache environment
}
\keyword{internal}
