% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\docType{data}
\name{safedata.env}
\alias{safedata.env}
\alias{index}
\title{Dataset index cache}
\format{An object of class \code{environment} of length 0.}
\usage{
safedata.env
}
\description{
The \code{safedata.env} environment within the package namespace is 
used to cache a copy of the dataset index, rather than needing to 
repeatedly read from file. This index is used internally by functions
through the internal \code{get_index} function and is not intended
for the end user.
}
\keyword{internal}
