% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safedata.R
\docType{package}
\name{safedata}
\alias{safedata}
\alias{safedata-package}
\title{safedata: Finding and using data from the SAFE Project.}
\description{
The SAFE Project is one of the largest ecological experiments in the world.
We are studying how biodiversity and ecosystem function change as forests are 
modified by human activities. As well as studying the change, we also
explore whether preserving sections of forest within modified landscapes
and around waterways can protect biodiversity and ecosystem function, 
and how much protection is needed to be effective.
}
\details{
The project is a multinational collaboration that has generated a wide range
of research data. We archive and document this data using the Zenodo file
repository and maintain spatial, taxonomic and other indices to support data
discovery and reuse. This package provides a range of tools to make it easier
to find and analyse our research data outputs.
}
\section{Links}{

   \url{https://www.safeproject.net/datasets/view_datasets}
   \url{https://www.zenodo.org/communities/safe}
}

