% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\docType{data}
\name{safedata_env}
\alias{safedata_env}
\alias{index}
\title{Index file memory cache}
\format{An object of class \code{environment} of length 0.}
\usage{
safedata_env
}
\description{
The \code{safedata_env} environment within the package namespace is
used to cache a copy of the dataset index, gazetteer and location aliases
rather than needing to repeatedly read these from file. This environment
is accessed by internal functions and is not intended for the end user.
}
\keyword{internal}
