\name{safer-package}
\alias{safer-package}
\alias{safer}
\docType{package}
\title{
\packageTitle{safer}
}
\description{
\packageDescription{safer}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{safer}
\packageIndices{safer}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{safer}

Maintainer: \packageMaintainer{safer}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
