% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchSAFE.R
\name{searchAuthor}
\alias{searchAuthor}
\title{Search SAFE projects by author name.}
\usage{
searchAuthor(author, ids = NULL, mostRecent = FALSE)
}
\arguments{
\item{author}{A character string used to search for datasets by author's 
full (or partial) names.}

\item{ids}{Numeric that restricts the returned API search to a subset of
Zenodo record IDs given in the \code{ids} parameter. This is typically
used to refine previous searches.}

\item{mostRecent}{Logical indicating whether to restrict the API to
returning only the most recent versions of the datasets found. By default
all versions of the discovered projects are returned.  #'}
}
\value{
An array of Zenodo record IDs for the SAFE project datasets found
  to match the search criteria.
}
\description{
Search for SAFE datasets by author name. \code{searchAuthor} enables users
to search the Zenodo database for SAFE projects containing the given
\code{author} among the author list. Partial author names may also be used.
Note that this function is not case sensitive.
}
\note{
This function returns Zenodo record IDs. On the Zenodo database each
  version of a SAFE project has a unique record ID. Different versions of
  the same dataset are grouped together under a single concept ID. See
  \url{https://help.zenodo.org/} for more information on Zenodo versioning.
}
\examples{
  searchAuthor('Ewers')
  searchAuthor('Ew')
  searchAuthor('Ewers', mostRecent = TRUE)
  searchAuthor('Ewers', ids = c(1303014, 1237725, 3081059, 1323504, 3251902))
}
\seealso{
\code{\link{getSafe}} for downloading SAFE datasets.
}
