% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchSAFE.R
\name{searchDates}
\alias{searchDates}
\title{Search SAFE projects by date.}
\usage{
searchDates(dates, matchType = "intersect", ids = NULL,
  mostRecent = FALSE)
}
\arguments{
\item{dates}{A character string containing one or two (comma separated) 
dates in ISO format (yyyy-mm-dd).}

\item{matchType}{A character string - one of 'intercept', 'contain', or 
'within' used to match the provided dates to the temporal extents of SAFE
datasets. Defaults to 'intersect'. The 'contain' option returns datasets
that span a date range, while 'within' returns datasets that fall within
the given range. Note that this parameter is only utilised when a date
range is specified (i.e. it is not used for single-date searches).}

\item{ids}{Numeric that restricts the returned API search to a subset of
Zenodo record IDs given in the \code{ids} parameter. This is typically
used to refine previous searches.}

\item{mostRecent}{Logical indicating whether to restrict the API to
returning only the most recent versions of the datasets found. By default
all versions of the discovered projects are returned.  #'}
}
\value{
An array of Zenodo record IDs for the SAFE project datasets found
  to match the search criteria.
}
\description{
Search for SAFE datasets by temporal extent. \code{searchDates} allows
users to search the Zenodo database for SAFE projects conducted on a
specific date, or over a given date range.
}
\note{
This function returns Zenodo record IDs. On the Zenodo database each
  version of a SAFE project has a unique record ID. Different versions of
  the same dataset are grouped together under a single concept ID. See
  \url{https://help.zenodo.org/} for more information on Zenodo versioning.
}
\examples{
  searchDates('2014-06-12')
  searchDates('2014-06-12,2015-06-11')
  searchDates('2014-06-12,2015-06-11', matchType = 'contain')
  searchDates('2014-06-12,2015-06-11', matchType = 'within')
  searchDates('2014-06-12,2015-06-11', mostRecent = TRUE)
  searchDates('2014-06-12,2015-06-11', 
              ids = c(3247465, 3081059, 1478526, 3251902))
}
\seealso{
\code{\link{getSafe}} for downloading SAFE datasets.
}
