% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchSAFE.R
\name{searchFields}
\alias{searchFields}
\title{Search SAFE projects by data field information.}
\usage{
searchFields(fieldText = NULL, fieldType = NULL, ids = NULL,
  mostRecent = FALSE)
}
\arguments{
\item{fieldText}{Character string to search for within the data field name
and description.}

\item{fieldType}{Character string to search for within the field type (for
example field type \code{numeric}).}

\item{ids}{Numeric that restricts the returned API search to a subset of
Zenodo record IDs given in the \code{ids} parameter. This is typically
used to refine previous searches.}

\item{mostRecent}{Logical indicating whether to restrict the API to
returning only the most recent versions of the datasets found. By default
all versions of the discovered projects are returned.}
}
\value{
An array of Zenodo record IDs for the SAFE project datasets found
  to match the search criteria.
}
\description{
Search for SAFE datasets by data field information. \code{searchFields}
allows users to search the Zenodo database for SAFE projects recording
specific data types, for example temperature data, or numeric data types.
}
\note{
This function returns Zenodo record IDs. On the Zenodo database each
  version of a SAFE project has a unique record ID. Different versions of
  the same dataset are grouped together under a single concept ID. See
  \url{https://help.zenodo.org/} for more information on Zenodo versioning.
}
\examples{
  searchFields(fieldText = 'temperature')
  searchFields(fieldType = 'numeric')
  searchFields(fieldText = 'temperature', fieldType = 'numeric')
  searchFields(fieldText = 'temperature', mostRecent = TRUE)
  searchFields(fieldText = 'temperature', ids = c(3238498, 3247485, 3251900))
}
\seealso{
\code{\link{getSafe}} for downloading SAFE datasets.
}
