% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchSAFE.R
\name{searchSpatial}
\alias{searchSpatial}
\title{Search SAFE projects by spatial sampling area/named location.}
\usage{
searchSpatial(wkt = NULL, location = NULL, distance = NULL,
  ids = NULL, mostRecent = FALSE)
}
\arguments{
\item{wkt}{Character string specifying a well-known text geometry. This is
assumed to use latitude and longitude in WGS84 (EPSG:4326), more
information on which can be read at
\url{https://en.wikipedia.org/wiki/World_Geodetic_System}. Useful
guidelines on constructing well-known text representations of geometry
can be found at
\url{https://en.wikipedia.org/wiki/Well-known_text_representation_of_geometry}.}

\item{location}{Character string giving a location name used to select
query geometry from the SAFE gazetteer, more information on which can be
found at \url{https://www.safeproject.net/info/gazetteer}.}

\item{distance}{Numeric search distance, specified in metres. All geometries
are converted to UTM 50N projection to provide appropriate distance
searching.}

\item{ids}{Numeric that restricts the returned API search to a subset of
Zenodo record IDs given in the \code{ids} parameter. This is typically
used to refine previous searches.}

\item{mostRecent}{Logical indicating whether to restrict the API to
returning only the most recent versions of the datasets found. By default
all versions of the discovered projects are returned.  #'}
}
\value{
An array of Zenodo record IDs for the SAFE project datasets found
  to match the search criteria.
}
\description{
Search for SAFE datasets on the Zenodo database using either a 
user-provided geometry or the geometry of a named location from the SAFE
gazetteer (\url{https://www.safeproject.net/info/gazetteer}). The sampling
locations provided in each SAFE dataset are tested to see if they intersect
the search geometry. A buffer \code{distance} can aso be provided to search
around the query geometry.
}
\details{
\code{searchSpatial} will not retrieve datasets that have not provided
sampling locations or use newly defined locations that are missing
coordinate information. When searching by bounding box (geometric shape)
the dataset geographic extent - which is provided for all datasets - is used.

Note that only one of \code{wkt} (well-known text geometry) or
\code{location} (named location on the SAFE gazetteer) should be provided.
}
\note{
This function returns Zenodo record IDs. On the Zenodo database each
  version of a SAFE project has a unique record ID. Different versions of
  the same dataset are grouped together under a single concept ID. See
  \url{https://help.zenodo.org/} for more information on Zenodo versioning.
  The function will not retrieve datasets for which no location information
  is provided.
}
\examples{
  searchSpatial(wkt = 'Point(116.5 4.75)')
  searchSpatial(wkt = 'Point(116.5 4.75)', distance = 100000)
  searchSpatial(wkt = 'Polygon((110 0, 110 10,120 10,120 0,110 0))')
  searchSpatial(location = 'A_1')
  searchSpatial(location = 'A_1', distance = 2500)
  searchSpatial(location = 'A_1', mostRecent = TRUE)
  searchSpatial(location = 'A_1', distance = 2500,
                ids = c(1303018, 1237730, 1198585, 1228188, 1198461, 119832))
}
\seealso{
\code{\link{getSafe}} for downloading SAFE datasets,
  \url{https://en.wikipedia.org/wiki/Well-known_text_representation_of_geometry}
  for setting up well-known text geometries,
  \url{https://www.safeproject.net/info/gazetteer} for an index of defined
  SAFE locations.
}
