% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchSAFE.R
\name{searchTaxa}
\alias{searchTaxa}
\title{Search SAFE projects by taxa.}
\usage{
searchTaxa(searchType, searchVal, ids = NULL, mostRecent = FALSE)
}
\arguments{
\item{searchType}{Character that specifies taxa search criteria, which
should be one of "name" (taxa name), "gbif_id" (ID number on the GBIF
database, \url{https://www.gbif.org/en/}), or "rank" (species, family, etc).}

\item{searchVal}{The value of parameter \code{searchType} to search against.}

\item{ids}{Numeric that restricts the returned API search to a subset of
Zenodo record IDs given in the \code{ids} parameter. This is typically
used to refine previous searches.}

\item{mostRecent}{Logical indicating whether to restrict the API to
returning only the most recent versions of the datasets found. By default
all versions of the discovered projects are returned.  #'}
}
\value{
An array of Zenodo record IDs for the SAFE project datasets found
  to match the search criteria.
}
\description{
Search for SAFE datasets by taxon information. \code{searchTaxa} allows
users to search the Zenodo database for SAFE projects containing data
pertaining to particular taxa. Taxa may be defined using one of three
\code{searchType} identifiers: name, gbif_id, or rank.
}
\note{
This function returns Zenodo record IDs. On the Zenodo database each
  version of a SAFE project has a unique record ID. Different versions of
  the same dataset are grouped together under a single concept ID. See
  \url{https://help.zenodo.org/} for more information on Zenodo versioning.
}
\examples{
  searchTaxa(searchType = 'name', searchVal = 'Formicidae')
  searchTaxa(searchType = 'gbif_id', searchVal = 4342)
  searchTaxa(searchType = 'rank', searchVal = 'Family')
  searchTaxa(searchType = 'name', searchVal = 'Formicidae', mostRecent = TRUE)
  searchTaxa(searchType = 'name', searchVal = 'Formicidae', 
             ids = c(1198302, 1198472, 1237732, 1198839, 1303014, 1237725))
}
\seealso{
\code{\link{getSafe}} for downloading SAFE datasets,
  \url{https://www.gbif.org/en/} for information on the GBIF database.
}
