% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchSAFE.R
\name{searchText}
\alias{searchText}
\title{Search SAFE projects by free text search.}
\usage{
searchText(text, ids = NULL, mostRecent = FALSE)
}
\arguments{
\item{text}{Character string to look for within a SAFE dataset, worksheet,
title, field description, and dataset keywords.}

\item{ids}{Numeric that restricts the returned API search to a subset of
Zenodo record IDs given in the \code{ids} parameter. This is typically
used to refine previous searches.}

\item{mostRecent}{Logical indicating whether to restrict the API to
returning only the most recent versions of the datasets found. By default
all versions of the discovered projects are returned.  #'}
}
\value{
An array of Zenodo record IDs for the SAFE project datasets found
  to match the search criteria.
}
\description{
Search for SAFE datasets using a free text search. \code{searchText} allows
users to perform a free text search on SAFE projects stored on the Zenodo
database. This function will search for values matching \code{text} within
dataset, worksheet, and data field descriptions, as well as in titles and
dataset keywords.
}
\note{
This function returns Zenodo record IDs. On the Zenodo database each
  version of a SAFE project has a unique record ID. Different versions of
  the same dataset are grouped together under a single concept ID. See
  \url{https://help.zenodo.org/} for more information on Zenodo versioning.
}
\examples{
  searchText('forest')
  searchText('ant')
  searchText('forest', mostRecent = TRUE)
  searchText('forest', ids = searchText('ant'), mostRecent = TRUE)
}
\seealso{
\code{\link{getSafe}} for downloading SAFE datasets.
}
