% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{set_example_safedata_dir}
\alias{set_example_safedata_dir}
\title{Functions to use an example data directory for package examples}
\usage{
set_example_safedata_dir(on = TRUE)
}
\value{
The set_example_safedata_dir function returns the path of the
   example directory invisibly (on = TRUE) or the path of any restored
   user data directory (on =FALSE).
}
\description{
The documentation of \code{safedata} includes code examples that require
a real data directory but package code must not write within the package
structure. The \code{set_example_safe_dir} function is used in code
examples to create a safedata directory using local resources in a
temporary directory. This example directory should only be created
once per session.
}
\details{
The argument \code{on=FALSE} can be used to restore any existing data
directory set by the user.
}
\seealso{
\code{\link{set_safedata_dir}}
}
