% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{set_safedata_paths}
\alias{set_safedata_paths}
\title{Sets the paths to the core index paths in a safedata directory.}
\usage{
set_safedata_paths(safedir)
}
\description{
Sets the paths to the core index paths in a safedata directory.
}
\keyword{internal}
