% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setSafeDir}
\alias{setSafeDir}
\title{Set the local SAFE data directory}
\usage{
setSafeDir(dir = NULL)
}
\arguments{
\item{dir, }{the directory to set as the SAFE_data_dir (defaults to 
\code{getwd()})}
}
\description{
Sets the local SAFE data directory using R "options". If an invalid/no
directory is specified, defaults to the current working directory.
}
