% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importSAFE.R
\name{simpleCap}
\alias{simpleCap}
\title{Capitalize the first letter of each word in a string}
\usage{
simpleCap(str)
}
\arguments{
\item{str}{The string to be capitalized}
}
\value{
The original string pasted together with all whitespace removed and
  the first letter of each word capitalized
}
\description{
This function takes a string, capitalizes the first letter of each word
within it, removes all whitespace, and combines the result into a single
word, returning the result.
}
\examples{
  simpleCap('The quick brown fox jumps over the lazy dog')
}
