% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSAFE.R
\name{updateIndex}
\alias{updateIndex}
\title{Update the SAFE index file}
\usage{
updateIndex(id, isConceptId = TRUE, dir = NULL)
}
\arguments{
\item{id}{the concept ID/record ID associated with the version of the SAFE
project being accessed}

\item{isConceptId}{boolean indicating whether parameter \code{ID} is a 
concept ID or a record ID. This changes the operation of the function}

\item{dir}{the directory into which the index.rds file should be stored,
which defaults to the \code{SAFE_data_dir}}
}
\description{
The SAFE index file contains version and file-download information for all
SAFE projects that have been requested by the user through the 
\code{safedata} package. The intended use of the index file is to provide
an offline cache of information associated with the varied records on the
SAFE Zenodo database to streamline operation of \code{safedata} and 
reduce redundancies. The index file is a dataframe stored (by default) in
the \code{SAFE_data_dir}. Records are grouped by concept ID. Meta 
information, including DOI, record ID, date created, ddataset embargo
status, open date, and local file download status (TRUE/FALSE) are then
saved for each record within the concept IDs requested. Every time a call
is made to the Zenodo database for a SAFE dataset, \code{updateIndex} runs,
either adding new version data or overwriting existing entries, depending
on whether the concept ID has been previously requested.
}
\details{
\code{updateIndex} takes a Zenodo ID as its main input, which may be a
concept ID (the default, \code{isConceptId = TRUE}), or the ID of a
specific record entry (\code{isConceptId = FALSE}). Under the default
behaviour, information pertaining to all versions of a particular concept
are accessed through a call to \code{\link{zenodoVersionsApiLookup}} and
stored within the index file. When a record ID is passed as input, it is
assumed that it is because the dataset associated with this version has
been downloaded and, after a directory check, the download status of this
entry is modified. At the end of each call the modified index dataframe is
saved in the local SAFE directory.

The SAFE index file is used downstream to identify and notify the user of
the most recent available versions of a project, or in the event that the
user is attempting to access a dataset that already exists within the
\code{SAFE_data_dir}.
}
\examples{
  # add version information for concept ID 3081058
  setSafeDir("C:/Users/User/SAFE_data/")
  updateIndex(3081058)
  
  # update download status of record 3081059 to TRUE
  updateIndex(3081059, isConceptId = FALSE)
}
\seealso{
\code{\link{zenodoVersionsApiLookup}}, \code{\link{setSafeDir}},
  \code{\link{buildVersionsDataframe}}
}
