% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSAFE.R
\name{updateVersionCache}
\alias{updateVersionCache}
\title{Update the offline version cache for a given SAFE project}
\usage{
updateVersionCache(versions, dir)
}
\arguments{
\item{versions}{the full versions list from the Zenodo 'versions' API call}

\item{dir}{local SAFE cache directory}
}
\value{
a dataframe containing access status information for each version
  of the specified SAFE project, which is saved at the \code{dir}.
}
\description{
Takes the output from a call to the Zenodo 'versions' API and processes the
returned \code{list} into a dataframe storing version details for all
records associated with a given SAFE project (identified using the concept
record ID number). Version fields include the DOI, record ID, creation
date, access status ('open', 'embargoed', or 'closed'), and - in the case 
of embargoed records - the open date (the date that the record will be
publicly accessible) for all records associated with the concept ID. The
constructed dataframe is stored at the specified directory \code{dir}.
}
\seealso{
\code{\link{zenodoVersionsApiLookup}}
}
