% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSAFE.R
\name{zenodoRecordApiLookup}
\alias{zenodoRecordApiLookup}
\title{Return information of the given id using the Zenodo record API}
\usage{
zenodoRecordApiLookup(id)
}
\arguments{
\item{id}{the data ID number, which in this instance should be either a
SAFE project DOI or concept ID number}
}
\value{
\code{list} object containing information from the Zenodo API call
}
\description{
This function returns a \code{list} object from an API call to the Zenodo
cloud database using the supplied ID number. The \code{list} object
contains varied record information for the ID.
}
\note{
Zenodo Record API: \url{https://zenodo.org/api/records/}
}
