% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSAFE.R
\name{zenodoVersionsApiLookup}
\alias{zenodoVersionsApiLookup}
\title{Perform a lookup on the Zenodo database with an API to access information
for all versions of the specified id.}
\usage{
zenodoVersionsApiLookup(id)
}
\arguments{
\item{id}{the data ID number, which in this instance should be a concept ID}
}
\value{
\code{list} object containing information on all versions of the ID
  from the Zenodo API call
}
\description{
This function returns a \code{list} onject from an API call to the Zenodo
cloud database using the supplied record ID number. The returned 
\code{list}contains information on all available versions of the data
assigned to that ID. Note that this API call only returns a populated 
\code{list} if a concept record ID is passed, otherwise the \code{list} is
empty.
}
\note{
Zenodo Versions API example: 
  \url{https://zenodo.org/api/records/?q=conceptrecid:1198692&all_versions&sort=-version&size=10&page=1}
}
