% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importSAFE.R
\name{addTaxonHeirarchies}
\alias{addTaxonHeirarchies}
\title{Add taxonomic heirarchies to SAFE object}
\usage{
addTaxonHeirarchies(safeObj, ...)
}
\arguments{
\item{safeObj}{An existing SAFE data object}

\item{...}{Optional arguments to pass to \code{\link[rgbif]{name_backbone}}}
}
\value{
The updated SAFE object \code{safeObj} with a dataframe of
  taxonomic heirarchies for all taxa listed in the Taxa worksheet of the
  data file. This is accessed using the reference \code{TaxaTree}
}
\description{
Adds a dataframe of taxonomic heirarchies for all taxa reported in the SAFE
Taxa worksheet to the supplied SAFE object. When no Taxa worksheet has been
supplied, this function flags a warning.
}
\note{
Not all SAFE data files contain the Taxa worksheet. In this case, a
  value of \code{NA} will be assigned to the \code{TaxaTree}
}
\seealso{
\code{\link{getNameBackbone}}
}
