% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa.R
\name{get_phylogeny}
\alias{get_phylogeny}
\title{Get a phylogeny for a dataset}
\usage{
get_phylogeny(record)
}
\arguments{
\item{record}{A single dataset record id}
}
\value{
An \code{ape} \code{phylo} object.
}
\description{
Get a phylogeny for a dataset
}
