% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa.R
\name{get_taxon_coverage}
\alias{get_taxon_coverage}
\title{Retrieve the current taxon index}
\usage{
get_taxon_coverage()
}
\value{
A data frame containing higher taxon level information for each taxon
  in the database taxon index.
}
\description{
This function downloads the current taxon index across published  datasets
and returns a formatted taxon table showing the taxonomic hierarchy of each
taxon. It requires an internet connection.
}
\details{
Note that the use of 'user' defined taxa can lead to some unusual entries. There
is nothing to stop a user defining 'Gallus gallus' as a 'user' taxon with Animalia
as a parent taxon.
}
\examples{
   \donttest{
    all_taxa <- get_taxon_coverage()
    }
}
\seealso{
\code{\link{get_taxa}}
}
