% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa.R
\name{get_taxon_graph}
\alias{get_taxon_graph}
\title{Get a graph of the taxa in a dataset}
\usage{
get_taxon_graph(record, which = c("gbif", "ncbi"))
}
\arguments{
\item{record}{A single dataset record id}

\item{which}{Return the GBIF or NCBI taxon graph}
}
\value{
An \code{\link[igraph:make_graph]{graph}} object.
}
\description{
This function loads the  taxon index for either of the NCBI or GBIF
taxa present in a dataset (see \code{\link{get_taxa}}) and creates a
taxonomic graph for the dataset using the \pkg{igraph} package.
}
\details{
The graph vertices are the different taxa used in the dataset along
with nodes for all parent taxa of those taxa. The graph vertex
names are arbitrary strings (tx1, ...) but the vertex attributes
contain the worksheet name, validated taxon name, taxon rank and taxon
status (see \code{names(vertex_attr(obj))}).
}
\examples{
   set_example_safe_dir()
   beetle_graph <- get_taxon_graph(1400562)
   plot(beetle_graph, vertex.label.cex = 0.6, vertex.size = 15,
        vertex.size2 = 3, vertex.shape = "rectangle")
   # show worksheet names for tips
   wsn <- igraph::vertex_attr(beetle_graph, "worksheet_name")
   txn <- igraph::vertex_attr(beetle_graph, "taxon_name")
   labels <- ifelse(is.na(wsn), txn, wsn)
   is_leaf <- igraph::vertex_attr(beetle_graph, "leaf")
   vert_col <- ifelse(is_leaf, "cornflowerblue", "grey")
   plot(beetle_graph, vertex.label.cex = 0.6, vertex.size = 15,
        vertex.size2 = 3, vertex.shape = "rectangle",
        vertex.label = labels, vertex.color= vert_col)
   unset_example_safe_dir()
}
