% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importSAFE.R
\name{importSafe}
\alias{importSafe}
\title{Import a SAFE data file}
\usage{
importSafe(filePath)
}
\arguments{
\item{filePath}{Full path to the .xlsx SAFE file}
}
\value{
A \code{safe_data} object with \code{Summary} information added
}
\description{
Create a new \code{safe_data} object with minimal summary information using
the SAFE dataset located at the specified \code{filePath}. Note that
currently only '.xlsx' file formats are supported - see
\url{https://safe-dataset-checker.readthedocs.io/en/latest/data_format/overview/#excel-format-overview}
for more information.
}
\seealso{
\code{\link{getSafe}} for downloading SAFE files from the Zenodo
  cloud database, \url{https://safe-dataset-checker.readthedocs.io/en/latest/data_format/overview/#excel-format-overview}
  for an overview on Excel file formats used for SAFE data submissions
}
