% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSAFE.R
\name{processSafe}
\alias{processSafe}
\title{Get record info for specified SAFE project ID from the Zenodo database}
\usage{
processSafe(id, dir, updateIndex = TRUE, overwrite = FALSE)
}
\arguments{
\item{id}{SAFE project record ID number. This could be a concept record
ID or a DOI identifier for a specific record. In the event that a concept
record ID is passed, the function will attempt to return the most recent
'open' dataset associated with the project}

\item{dir}{directory into which data will be stored}

\item{updateIndex}{\code{boolean}, should the SAFE index file be updated?}

\item{overwrite}{Should existing copies of the data be overwritten.}
}
\value{
\code{list} object containing information from the Zenodo API call
}
\description{
This function returns a \code{list} object from an API call to the Zenodo
database using the supplied SAFE project ID number. The API call contains
information required to download the requested SAFE dataset into the
specified directory.
}
\details{
The function will automatically create a data folder at the directory
\code{dir} and (optionally, \code{updateIndex} boolean) updates the SAFE
index offline cache - a table that records version and file information for
SAFE datasets that have been requested by the user. The index cache is
automatically downloaded into the \code{dir} the first time a record from
a particular concept ID is accessed.

\code{processSafe} additionally performs a series of checks on the record.
Specifically, an error is raised in the event that the ID does not exist on
Zenodo, or is not located in the SAFE community, or if the record is
unavailable (i.e. is 'embargoed' or 'closed'). A warning message will be
presented if a more recent version of the requested record is available for
download.
}
\seealso{
\code{\link{zenodoRecordApiLookup}}, \code{\link{updateIndex}}
}
